Array Pairs solution reworked!
Previous solution was moved to solution-working-but-ugly.js

Yet another solution-working-array.js was created. It was an attempt to swap arrays and maps to win on removing not needed values by reducing length of sorted array. However, it did not improve the solution that much.

Final solution was created basing on idea to build the tree first and then to move from leaves to root to simplify calculation of counters. It cut 1 second from calculation time on worst test case (014). Also, I've reworked the structure of solution: it was broken down to functions so it is simpler to read the code.