Burger Happiness: solved small cases but failed with big ones.
Note: debug info is kept in the solution for a while since I need to improve it. It helps to find a mistake in calculations.

It is fast but eats N ** 2 memory what cause an exception in case of gin N.

BCounter is added into every Happiness and actually cost N size right now. I need to rework BCounter to something else than Bit-based solution to reduce its memory cost but keep it fast as log(n) still. Current idea is to use AVL BST with tracking sums of left subtrees. Will see if it works.