Burger Happiness: solved
It appeared that very first solution I was able to figure out is the right one. It is really strange because I can prove that in worst case it is N ** 2. It is required to make a test with 100000 restaurants and make every B small enough so that every intermediate max happiness should be propagated across all 100000 elements. In "faster" solutions I've tried to implement some magic to convert this solution into N * log(N) (used tree and "annihilation" functions based on (a & -a) magic... and it appeared that it is not necessary and brut force solution is just fine enough! Surprise!